Messages and I2C module.
Изменил механизм отправки сообщений (теперь есть два параметра). Соответственно отредактировал код под новые сообщения. Также сделал отдельными файлами модуль I2C, чтобы возможно было прикрутить не только дисплей.
