+++ My Libraries/Code/inc/messages.h
@@ -12,11 +12,14 @@
 
 /*************************	Define	*************************/
 
-#define MaxNumbMessages		30
+#define MaxNumbMessages			30
 
-#define LedOnMsg			1
+#define LedOnMsg				1
 
+#define I2C1StartTransaction	10					// Начать передачу
+#define I2CAddrOk				11					// Есть ответ от устройства с заданным адресом
 
+#define LCDStartInit			13					// Начать инициализацию дисплея
 
 #define ModbusError				21					// Ошибка Modbus
 #define ModbusReciveSymbol		22					// USART-Modbus принял символ
@@ -26,15 +29,10 @@
 #define ModbusCRCNotOk			26					// Несовпадение контрольной суммы
 #define ModbusMessageReceived	27					// Сообщение получено
 
-#define I2C1StartTransaction	10					// Начать передачу
-#define LCDStartInit			11					// Начать инициализацию дисплея
-
-
-
 /*************************	 Code	*************************/
 
 void InitMessage (void);
-void SendMessage (int Msg);
+void SendMessage (uint16_t Message, uint16_t ParametrOne, uint16_t ParametrTwo);
 void ProcessMessage (void);
-uint8_t GetMessage (int Msg);
+uint8_t GetMessage (uint8_t Message);
