+++ My Libraries/Code/src/lcd16xx.c
@@ -32,160 +32,10 @@ const uint8_t BuferLCDInit[] = 	{0x3C, 0x38,  									/*	 0 - 1	-> пауза 6
 const uint8_t DDRAMLCD1602 [2][16] = {{0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F},
 						  	  	   	  {0x40, 0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, 0x49, 0x4A, 0x4B, 0x4C, 0x4D, 0x4E, 0x4F}};
 
-uint8_t I2C1Data [I2C1DataBuferLenght];
-
-uint8_t i2cSendStates;
-uint8_t _i2cSendStates;
-uint8_t i2cEntry;
-
-uint8_t I2C1NumberOfTransaction;						//	количество передач по I2C
-uint8_t I2C1SendBuferLenght;							//	длина перевадаемого буфера
-
 uint8_t lcdStates;
 uint8_t _lcdStates;
 uint8_t lcdEntry;
 
-/*************************	 Code	*************************/
-
-void InitI2C1 (void){
-
-/* I2C1 GPIO Configuration */
-
-	GPIOB->MODER &= ~GPIO_MODER_MODER6;
-	GPIOB->MODER |= GPIO_MODER_MODER6_1;				// Configuring the I / O Direction Mode for the Alternate Function
-	GPIOB->OTYPER |= GPIO_OTYPER_OT6;					// Configuring the output type of the open-drain
-	GPIOB->OSPEEDR &= ~GPIO_OSPEEDR_OSPEED6;
-	GPIOB->OSPEEDR |= GPIO_OSPEEDR_OSPEED6_0;			// Configuring the output speed is low
-	GPIOB->PUPDR &= ~GPIO_PUPDR_PUPD6;
-	GPIOB->PUPDR |= GPIO_PUPDR_PUPD6_0;					// Configuring the I / O pull-up
-	GPIOB->AFR[0] &= ~GPIO_AFRL_AFRL6;					// Enable AF4 (I2C1...3) on port PB6
-	GPIOB->AFR[0] |= GPIO_AFRL_AFRL6_2;
-
-	GPIOB->MODER &= ~GPIO_MODER_MODER7;
-	GPIOB->MODER |= GPIO_MODER_MODER7_1;				// Configuring the I / O Direction Mode for the Alternate Function
-	GPIOB->OTYPER |= GPIO_OTYPER_OT7;					// Configuring the output type of the open-drain
-	GPIOB->OSPEEDR &= ~GPIO_OSPEEDR_OSPEED7;
-	GPIOB->OSPEEDR |= GPIO_OSPEEDR_OSPEED7_0;			// Configuring the output speed is low
-	GPIOB->PUPDR &= ~GPIO_PUPDR_PUPD7;
-	GPIOB->PUPDR |= GPIO_PUPDR_PUPD7_0;					// Configuring the I / O pull-up
-	GPIOB->AFR[0] &= ~GPIO_AFRL_AFRL7;					// Enable AF4 (I2C1...3) on port PB6
-	GPIOB->AFR[0] |= GPIO_AFRL_AFRL7_2;
-
-	RCC->APB1ENR |= RCC_APB1ENR_I2C1EN;					// Enable I2C1 clocking
-
-	I2C1->OAR2 &= ~I2C_OAR2_ENDUAL;						// Disable Dual addressing mode
-
-	I2C1->CR1 &= ~I2C_CR1_ENGC;							// Disable General call
-
-	I2C1->CR1 &= ~I2C_CR1_NOSTRETCH;					// Enable clock stretching
-
-	I2C1->CR1 &= ~I2C_CR1_SMBUS;						// Enable I2C mode
-
-	I2C1->CR2 &= ~I2C_CR2_FREQ;
-	I2C1->CR2 |= 48;									// Set 48 MHz frequency
-
-	I2C1->CCR &= ~I2C_CCR_FS;							// Sm mode I2C
-
-	I2C1->CCR &= ~I2C_CCR_CCR;
-	I2C1->CCR |= 240;
-
-	I2C1->TRISE = 3;
-
-	I2C1->CR2 |= I2C_CR2_ITEVTEN;						// Разрешаем прерывания по событиям отравки и др. (Event interrupt enable)
-
-	I2C1->CR1 |= I2C_CR1_PE;							// I2C1 Enable
-
-}
-
-void InitI2C1FSM (void){
-
-	InitI2C1();
-
-	i2cSendStates = 0;
-	_i2cSendStates = 0;
-	I2C1NumberOfTransaction = 0;
-
-	for (uint8_t i = 0; i < I2C1DataBuferLenght; i++){
-		I2C1Data[i] = 0x00;
-	}
-}
-
-void ProcessI2CWriteFSM (void){
-
-	if ( i2cSendStates != _i2cSendStates) i2cEntry = 1; else i2cEntry = 0;
-
-	_i2cSendStates = i2cSendStates;
-
-	switch (i2cSendStates){
-
-	case 0:
-
-		if (GetMessage(I2C1StartTransaction)){
-			if (I2C1SendBuferLenght != 0){
-				i2cSendStates = 1;
-			}
-		}
-		break;
-
-	case 1:
-
-		if (i2cEntry == 1){
-			I2C1->CR1 |= I2C_CR1_START;
-		}
-
-		if (GetGTimerVal(I2C1Timer) >= 1000){
-			i2cSendStates = 4;
-		}
-		break;
-
-	case 2:
-		if (i2cEntry == 1) StopGTimer(I2C1Timer);
-		break;
-
-	case 3:
-
-		if (I2C1NumberOfTransaction == I2C1SendBuferLenght){
-			i2cSendStates = 0;
-		}
-		else {
-			i2cSendStates = 1;
-		}
-		break;
-
-	case 4:
-		GPIOC->BSRR |= GPIO_BSRR_BS15;
-
-		break;
-	}
-
-}
-
-
-
-/*************************	 IRQ_Handler (Обработчики прерываний)	*************************/
-
-void I2C1_EV_IRQHandler (void){
-
-	if (I2C1->SR1 & I2C_SR1_SB){
-		(void) I2C1->SR1;
-		StartGTimer(I2C1Timer);
-		I2C1->DR = AddrDevice;
-	}
-
-	if (I2C1->SR1 & I2C_SR1_ADDR){
-		(void) I2C1->SR1;
-		(void) I2C1->SR2;
-		i2cSendStates = 2;
-		I2C1->DR = I2C1Data[I2C1NumberOfTransaction];
-	}
-
-	if (I2C1->SR1 & I2C_SR1_BTF){
-		I2C1->CR1 |= I2C_CR1_STOP;
-		I2C1NumberOfTransaction++;
-		i2cSendStates = 3;
-	}
-
-}
 
 /*************************	 Функции для дисплея 16xx (16xx Display function)	*************************/
 
@@ -205,9 +55,7 @@ void ProcessLcdFSM (void){
 	case 0:
 		if (GetMessage(LCDStartInit)){
 
-			for (uint8_t i = 0; i < 8; i++){
-				I2C1Data[i] = BuferLCDInit[i];
-			}
+			ClearI2C1DataBufer();
 			lcdStates = 1;
 		}
 		break;
@@ -215,48 +63,15 @@ void ProcessLcdFSM (void){
 	case 1:
 
 		if (lcdEntry == 1){
-			I2C1SendBuferLenght = 4;
-			SendMessage(I2C1StartTransaction);
-		}
-
-		if (I2C1NumberOfTransaction == 4){
-			StartGTimer(LCDTimer);
-			I2C1SendBuferLenght = 6;
-		}
-
-		if (GetGTimerVal(LCDTimer) > 10){
-			if (I2C1NumberOfTransaction == 4){
-				StopGTimer(LCDTimer);
-				SendMessage(I2C1StartTransaction);
-				StartGTimer(LCDTimer);
-			}
-			if (I2C1NumberOfTransaction == 6){
-				StopGTimer(LCDTimer);
-				I2C1SendBuferLenght = 8;
-				SendMessage(I2C1StartTransaction);
-			}
-			if (I2C1NumberOfTransaction == 8){
-				lcdStates = 2;
-			}
+			uint16_t LenghtPackage = 4;
+			uint16_t StartAdressPackage = 0;
+			SendMessage(I2C1StartTransaction, LenghtPackage, StartAdressPackage);
 		}
 		break;
 
 	case 2:
 
-		if (lcdEntry == 1){
-			for (uint8_t i = 0; i < 24; i++){
-				I2C1Data[i] = BuferLCDInit[i+8];
-			}
-			I2C1NumberOfTransaction = 0;
-			I2C1SendBuferLenght = 24;
-			SendMessage(I2C1StartTransaction);
-		}
 
-		if (I2C1NumberOfTransaction == 24){
-			I2C1NumberOfTransaction = 0;
-			I2C1SendBuferLenght = 0;
-			lcdStates = 3;
-		}
 		break;
 
 	case 3:
