+++ My Libraries/Code/src/virtualTimers.c
@@ -57,7 +57,7 @@ void TIM1_UP_TIM10_IRQHandler (void){
 		MRTUcount++;
 
 		if (MRTUcount >= 5){
-			SendMessage(ModbusRTUTimeOut);
+			SendMessage(ModbusRTUTimeOut, 0, 0);
 			ReceptionStatus = ReceptionStopped;
 		}
