+++ My Libraries/Debug/Code/src/main.d
@@ -15,6 +15,7 @@ Code/src/main.o: ../Code/src/main.c \
  C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/CMSIS/inc/core_cm4.h \
  C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/CRC16.h \
  C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/lcd16xx.h \
+ C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/I2C.h \
  C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/ledFSM.h
 
 C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/main.h:
@@ -48,5 +49,7 @@ C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/CMSIS/inc/core_cm4.h:
 C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/CRC16.h:
 
 C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/lcd16xx.h:
+
+C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/I2C.h:
 
 C:/Users/Human/Documents/GitHub/repository2/My\ Libraries/Code/inc/ledFSM.h:
