+++ My Libraries/Debug/objects.list
@@ -1,5 +1,6 @@
 "./CMSIS/src/system_stm32f4xx.o"
 "./Code/src/CRC16.o"
+"./Code/src/I2C.o"
 "./Code/src/lcd16xx.o"
 "./Code/src/ledFSM.o"
 "./Code/src/main.o"
